Refactor: A gigantic refactoring
Removed the usage of ECS container:
We now handle the file processors ourself in the runner class

Added more phpstan rules:
We now use a lot more rules from phpstan

Fixed a bug with non autoloaded files:
Files which is not part of the autoload did not work with phpstan.

Removed the need for phpstan analyser:
We now only use the parser, this improves the speed a lot

Merged sniff insight and sniff decorator together:
The sniff insights and decorator are now 1 class instead of two

Removed errorAndDiffCollector:
We don't use the collector anymore. We now store the sniff errors
inside the sniff itself.