Removed unneeded variable $testPreset from resolve method.
Rewrote guess method to now optionally accept a preset name to attempt to resolve. This is a breaking change since the return value will not be the preset name, but rather the full qualified class name of the preset.

Reworked the resolvePreset method. It should now return a valid custom preset first, then guess the preset, and finally if guessing fails it will return the default preset.