Updated guess method to now return the class name for the Preset instend of a human readable name.
Updated the resolve preset method to accept the config array as its first element, and also to return the DefaultPreset if not ot
her preset is found.

Updated the ConfigResolverTest to now use the preset's full qualified class name instead of a human readable name. Also updated t
he unknoww preset test to check for the DefaultPreset class instead of the InvalidPreset exception.

Removed InvalidPreset and PresetNotFound exception since neither was being used.