Fix broken test
This test's name implies parallel-lint should be run on `__DIR__`, but
in reality it's exactly the same command as the single-file test because
as long as only one file is in the collector, it will only scan that.

This updates the analyser to pass *two* files instead, thus forcing
SyntaxCheck to call parallel-lint with the common path instead of a
single filepath. Since it now scans the whole folder as the test
implies, we also sort the details list by comparing the file paths
otherwise the order would change on every test run.