Set working directory to fix --exclude args
The call to parallel-lint would always use the common path in tests and
any other time the common path differed from `getcwd()`. This meant any
--exclude args in `$cmdLine` would require fully resolved paths, which
would likely only end up with similar "arg list too long" issues.

Instead, this sets the working directory of the process when it doesn't
match the common path (and it's not a single file being analysed).
Behaviour is the same, but now parallel-lint can resolve the excludes.